feat: create the login in fake backend
To start develop the Front-End features was developed the simple
fake Back-End to make most easy of return the data in Front-End
and simplfy to code the validations, of inputs and data.

This commit introduces a new router in `/api/login` as a *POST* method,
sending in body email and password of user, and then validate if the
email of user and password contain in a database.json file.